requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
django-phonenumber-field==3.0.1
gunicorn==20.0.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
numpy==1.17.4
packaging==19.2
phonenumbers==8.10.22
Pillow==6.2.1
pipenv==2018.11.26
psycopg2==2.8.3
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
typed-ast==1.4.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.11.2
